Restructure into own namespace and multiple files
